# docs/openapi.yaml
openapi: 3.0.3
info:
  title: qf Control Plane API
  description: |
    REST API for the qf host-based firewall control plane.

    **Authentication**: Most endpoints require either a JWT cookie (`qf_token`, set by `POST /auth/login`)
    or an `Authorization: Bearer <token>` header (JWT or opaque API token).

    **Tenant**: Protected endpoints require the `X-Tenant-ID` header (UUID of the tenant).
  version: "0.5.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: qf_token
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant UUID
      schema:
        type: string
        format: uuid
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 500
        maximum: 1000
    StartTime:
      name: start
      in: query
      schema:
        type: string
        format: date-time
    EndTime:
      name: end
      in: query
      schema:
        type: string
        format: date-time
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Hosts ──────────────────────────────────────────────────────────────
    Host:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        hostname:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          enum: [pending, enrolling, active, degraded, offline]
        current_generation:
          type: integer
        last_heartbeat_at:
          type: string
          format: date-time
          nullable: true
        agent_version:
          type: string
          nullable: true
        kernel_version:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateHostRequest:
      type: object
      required: [hostname]
      properties:
        hostname:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          default: pending
...
    Policy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
        selector:
          type: object
          description: |
            Label-selector to match hosts. Example: `{"matchLabels": {"env": "prod"}}`.
            Empty selector matches all hosts.
        current_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePolicyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          default: 0
        selector:
          type: object
...
    ObjectGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [ipset, portset, hostset]
        name:
          type: string
        spec:
          type: object
          description: |
            Group definition. Examples:
            - ipset: `{"cidrs": ["10.0.0.0/8", "192.168.1.0/24"]}`
            - portset: `{"ports": [80, 443, 8080]}`
            - hostset: `{"hostnames": ["db.internal", "redis.internal"]}`
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateObjectGroupRequest:
      type: object
      required: [type, name]
      properties:
        type:
          type: string
          enum: [ipset, portset, hostset]
        name:
          type: string
        spec:
          type: object
    UpdateObjectGroupRequest:
      type: object
      required: [spec]
      properties:
        spec:
          type: object

    # ── Tokens ─────────────────────────────────────────────────────────────
    Token:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [single_host, bulk]
        target_host_id:
          type: string
          format: uuid
          nullable: true
        label_template:
          type: object
          additionalProperties:
            type: string
        max_uses:
          type: integer
        uses_count:
          type: integer
        expires_at:
          type: string
          format: date-time
        token:
          type: string
          description: Plain token value — only present in create response
    CreateTokenRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [single_host, bulk]
        target_host_id:
          type: string
          format: uuid
          description: Required when type=single_host
        label_template:
          type: object
          additionalProperties:
            type: string
        ttl_seconds:
          type: integer
          default: 3600
        max_uses:
          type: integer
          default: 1

    # ── Default Policy ─────────────────────────────────────────────────────
    DefaultPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        default_ingress_action:
          type: string
          enum: [allow, deny]
        default_egress_action:
          type: string
          enum: [allow, deny]
        updated_at:
          type: string
          format: date-time
    PutDefaultPolicyRequest:
      type: object
      properties:
        default_ingress_action:
          type: string
          enum: [allow, deny]
          default: deny
        default_egress_action:
          type: string
          enum: [allow, deny]
          default: deny

    # ── Events ─────────────────────────────────────────────────────────────
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        actor_type:
          type: string
        actor_id:
          type: string
          format: uuid
        action:
          type: string
        object_type:
          type: string
        object_id:
          type: string
          format: uuid
        before:
          type: object
          nullable: true
        after:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Auth ───────────────────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    MeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        tenant_id:
          type: string
          format: uuid

    # ── Users ──────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        status:
          type: string
          enum: [active, suspended]
        role:
          type: string
          enum: [admin, operator, auditor]
        is_oidc:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          type: string
          enum: [admin, operator, auditor]
...
    APIToken:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, auditor]
        created_by:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        token:
          type: string
          description: Plain token — only present in create response
    CreateAPITokenRequest:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, auditor]
        expires_at:
          type: string
          format: date-time
          nullable: true

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful; sets `qf_token` and `qf_refresh` cookies
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh cookie
      security: []
      responses:
        "200":
          description: New access token cookie set
        "401":
          description: Invalid or expired refresh token
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user identity
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC callback — exchange code for tokens
      security: []
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to /app after successful login
        "400":
          description: Invalid state or code

  # ── Hosts ─────────────────────────────────────────────────────────────────
  /hosts:
    get:
      tags: [Hosts]
      summary: List all hosts
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: List of hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Host'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Hosts]
      summary: Create a host
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHostRequest'
      responses:
        "201":
          description: Host created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /policies:
    get:
      tags: [Policies]
      summary: List policies
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Policy'
    post:
      tags: [Policies]
      summary: Create a policy
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
...
  /objectgroups:
    get:
      tags: [ObjectGroups]
      summary: List object groups
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - name: type
          in: query
          schema:
            type: string
            enum: [ipset, portset, hostset]
      responses:
        "200":
          description: Object groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectGroup'
    post:
      tags: [ObjectGroups]
      summary: Create an object group
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObjectGroupRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectGroup'
  /objectgroups/{id}:
    get:
      tags: [ObjectGroups]
      summary: Get an object group
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ResourceID'
      responses:
        "200":
          description: Object group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectGroup'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags: [ObjectGroups]
      summary: Update object group spec
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateObjectGroupRequest'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectGroup'
    delete:
      tags: [ObjectGroups]
      summary: Delete an object group
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ResourceID'
      responses:
        "204":
          description: Deleted

  # ── Tokens ────────────────────────────────────────────────────────────────
  /tokens:
    get:
      tags: [Tokens]
      summary: List enrollment tokens
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Token'
    post:
      tags: [Tokens]
      summary: Create an enrollment token
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        "201":
          description: Token created; plain token in `token` field (shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
...
  /default-policy:
    get:
      tags: [DefaultPolicy]
      summary: Get tenant default ingress/egress actions
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: Default policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultPolicy'
    put:
      tags: [DefaultPolicy]
      summary: Set tenant default ingress/egress actions
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutDefaultPolicyRequest'
      responses:
        "200":
          description: Updated default policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultPolicy'

  # ── Audit Log ─────────────────────────────────────────────────────────────
  /audit-log:
    get:
      tags: [AuditLog]
      summary: List audit log entries
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/StartTime'
        - $ref: '#/components/parameters/EndTime'
        - $ref: '#/components/parameters/Limit'
        - name: actor_type
          in: query
          schema:
            type: string
        - name: actor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: object_type
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'

  # ── Users ─────────────────────────────────────────────────────────────────
  /users:
    get:
      tags: [Users]
      summary: List users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create a user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          $ref: '#/components/responses/Forbidden'
...
  /api-tokens:
    get:
      tags: [APITokens]
      summary: List API tokens
      responses:
        "200":
          description: API tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIToken'
    post:
      tags: [APITokens]
      summary: Create an API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPITokenRequest'
      responses:
        "201":
          description: Token created; plain token in `token` field (shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIToken'
...